pyproject.toml
[tool.poetry]
name = "genie"
version = "0.1.0"
description = "Genie is an open-source levelling and rewards system built for the Discord community, inspired by Old School RuneScape."
authors = ["Tarran Prior <Tarranprior@protonmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.5.2"
python-dotenv = "^1.1.0"
loguru = "^0.7.3"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
